Improve sync performances
