+++ packages/twenty-server/src/engine/workspace-manager/workspace-migration-runner/workspace-migration-runner.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 
 import { WorkspaceMigrationModule } from 'src/engine/metadata-modules/workspace-migration/workspace-migration.module';
 import { WorkspaceDataSourceModule } from 'src/engine/workspace-datasource/workspace-datasource.module';
+import { WorkspaceMigrationColumnService } from 'src/engine/workspace-manager/workspace-migration-runner/services/workspace-migration-column.service';
 import { WorkspaceMigrationEnumService } from 'src/engine/workspace-manager/workspace-migration-runner/services/workspace-migration-enum.service';
 
 import { WorkspaceMigrationRunnerService } from './workspace-migration-runner.service';
@@ -14,6 +15,7 @@ import { WorkspaceMigrationTypeService } from './services/workspace-migration-ty
     WorkspaceMigrationRunnerService,
     WorkspaceMigrationEnumService,
     WorkspaceMigrationTypeService,
+    WorkspaceMigrationColumnService,
   ],
   exports: [WorkspaceMigrationRunnerService],
 })
